+++ src/SCEMonteCarlo.jl
@@ -73,6 +73,7 @@ public resolve_kt
 public ScaledTerm, SpinConfig, site_index, site_atom
 public site_coeffs!, delta_energy, site_gradient, energy_gradient, energy_gradient!
 public philox_block, philox_normal2
+public model_fingerprint
 public LogBinner, BinStore, jackknife, std_error, tau_int, bin_means
 public ChainState, SweepScratch, metropolis_sweep!, overrelaxation_sweep!
 public to_matrix, from_matrix